+++ sdk/appcontainers/arm-appcontainers/src/models/mappers.ts
@@ -872,6 +872,13 @@ export const Login: coreClient.CompositeMapper = {
           className: "LoginRoutes"
         }
       },
+      tokenStore: {
+        serializedName: "tokenStore",
+        type: {
+          name: "Composite",
+          className: "TokenStore"
+        }
+      },
       preserveUrlFragmentsForLogins: {
         serializedName: "preserveUrlFragmentsForLogins",
         type: {
@@ -914,6 +921,50 @@ export const LoginRoutes: coreClient.CompositeMapper = {
     modelProperties: {
       logoutEndpoint: {
         serializedName: "logoutEndpoint",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const TokenStore: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "TokenStore",
+    modelProperties: {
+      enabled: {
+        serializedName: "enabled",
+        type: {
+          name: "Boolean"
+        }
+      },
+      tokenRefreshExtensionHours: {
+        serializedName: "tokenRefreshExtensionHours",
+        type: {
+          name: "Number"
+        }
+      },
+      azureBlobStorage: {
+        serializedName: "azureBlobStorage",
+        type: {
+          name: "Composite",
+          className: "BlobStorageTokenStore"
+        }
+      }
+    }
+  }
+};
+
+export const BlobStorageTokenStore: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "BlobStorageTokenStore",
+    modelProperties: {
+      sasUrlSettingName: {
+        serializedName: "sasUrlSettingName",
+        required: true,
         type: {
           name: "String"
         }
@@ -1029,6 +1080,27 @@ export const ForwardProxy: coreClient.CompositeMapper = {
       },
       customProtoHeaderName: {
         serializedName: "customProtoHeaderName",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const EncryptionSettings: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "EncryptionSettings",
+    modelProperties: {
+      containerAppAuthEncryptionSecretName: {
+        serializedName: "containerAppAuthEncryptionSecretName",
+        type: {
+          name: "String"
+        }
+      },
+      containerAppAuthSigningSecretName: {
+        serializedName: "containerAppAuthSigningSecretName",
         type: {
           name: "String"
         }
@@ -1692,6 +1764,12 @@ export const CertificateProperties: coreClient.CompositeMapper = {
         type: {
           name: "String"
         }
+      },
+      type: {
+        serializedName: "type",
+        type: {
+          name: "String"
+        }
       }
     }
   }
@@ -2176,6 +2254,18 @@ export const Ingress: coreClient.CompositeMapper = {
           name: "Composite",
           className: "CorsPolicy"
         }
+      },
+      additionalPortMappings: {
+        serializedName: "additionalPortMappings",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "IngressPortMapping"
+            }
+          }
+        }
       }
     }
   }
@@ -2354,6 +2444,35 @@ export const CorsPolicy: coreClient.CompositeMapper = {
         serializedName: "allowCredentials",
         type: {
           name: "Boolean"
+        }
+      }
+    }
+  }
+};
+
+export const IngressPortMapping: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "IngressPortMapping",
+    modelProperties: {
+      external: {
+        serializedName: "external",
+        required: true,
+        type: {
+          name: "Boolean"
+        }
+      },
+      targetPort: {
+        serializedName: "targetPort",
+        required: true,
+        type: {
+          name: "Number"
+        }
+      },
+      exposedPort: {
+        serializedName: "exposedPort",
+        type: {
+          name: "Number"
         }
       }
     }
@@ -5144,6 +5263,92 @@ export const AzureCredentials: coreClient.CompositeMapper = {
   }
 };
 
+export const ListUsagesResult: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ListUsagesResult",
+    modelProperties: {
+      value: {
+        serializedName: "value",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "Usage"
+            }
+          }
+        }
+      },
+      nextLink: {
+        serializedName: "nextLink",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const Usage: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "Usage",
+    modelProperties: {
+      unit: {
+        defaultValue: "Count",
+        isConstant: true,
+        serializedName: "unit",
+        type: {
+          name: "String"
+        }
+      },
+      currentValue: {
+        serializedName: "currentValue",
+        required: true,
+        type: {
+          name: "Number"
+        }
+      },
+      limit: {
+        serializedName: "limit",
+        required: true,
+        type: {
+          name: "Number"
+        }
+      },
+      name: {
+        serializedName: "name",
+        type: {
+          name: "Composite",
+          className: "UsageName"
+        }
+      }
+    }
+  }
+};
+
+export const UsageName: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "UsageName",
+    modelProperties: {
+      value: {
+        serializedName: "value",
+        type: {
+          name: "String"
+        }
+      },
+      localizedValue: {
+        serializedName: "localizedValue",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
 export const JobExecutionNamesCollection: coreClient.CompositeMapper = {
   type: {
     name: "Composite",
@@ -5271,6 +5476,13 @@ export const AuthConfig: coreClient.CompositeMapper = {
         type: {
           name: "Composite",
           className: "HttpSettings"
+        }
+      },
+      encryptionSettings: {
+        serializedName: "properties.encryptionSettings",
+        type: {
+          name: "Composite",
+          className: "EncryptionSettings"
         }
       }
     }